requirements.txt
asgiref==3.10.0
Django==5.2.7
django-ckeditor==6.7.3
django-environ==0.12.0
django-js-asset==3.1.2
et_xmlfile==2.0.0
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
pillow==11.3.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
